% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_mdd.R
\name{get_mdd}
\alias{get_mdd}
\title{Function to return mammal shapefiles of Rodentia order}
\source{
Expert range maps of global mammal distributions harmonized to
according mdd database from American Mammalogist society.
Data storage in \doi{doi:10.5281/zenodo.10806734}
}
\usage{
get_mdd(dir = NULL, version = "v1_2")
}
\arguments{
\item{dir}{A directory where to write the output}

\item{version}{The mdd version order want to download}
}
\value{
A SpatVector object with mammal shapefiles
}
\description{
Function to return mammal shapefiles of Rodentia order
}
\examples{
get_mdd(version = "test")
}
