% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_mdd_list.R
\name{get_mdd_list}
\alias{get_mdd_list}
\title{Function to return mammal species list avaliable by version}
\source{
Expert range maps of global mammal distributions harmonized to
three taxonomic authorities \doi{doi:10.5281/zenodo.10806734}
}
\usage{
get_mdd_list(version = "v1_2")
}
\arguments{
\item{version}{The taxonomy version}
}
\value{
A data.frame object with mammal species by version
}
\description{
Function to return mammal species list avaliable by version
}
\examples{
get_mdd_list(version = "test")
}
