% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_mdd_map.R
\name{get_mdd_map}
\alias{get_mdd_map}
\title{Title}
\source{
Expert curated range maps of global mammal  distributions
aligned according mdd database of American Mamalogist Society.
\doi{doi:10.5281/zenodo.10806734}
}
\usage{
get_mdd_map(species = NULL, order = NULL, version = "v1_3")
}
\arguments{
\item{species}{A vector of mammal species name.}

\item{order}{A vector of mammal orders to retrive.}

\item{version}{The version number of the mammal diversity database.}
}
\value{
A SpatVector object with mammal shapefiles
}
\description{
Title
}
\examples{
get_mdd_map(version = "test")
}
