% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/globals.R
\name{globalVariables}
\alias{globalVariables}
\title{Global Variables}
\arguments{
\item{x}{A string with the url to check if exist}
}
\value{
A logic value if url exists
}
\description{
Global Variables
}
\keyword{internal}
