% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mddSpList_v1_3.R
\docType{data}
\name{mddSpList_v1_3}
\alias{mddSpList_v1_3}
\title{Species list from Mammal Diversity Database}
\format{
\subsection{\code{mddSpList}}{

A data frame with 6513 rows and 8 columns:
\describe{
\item{Order}{Order of the species}
\item{Family}{Family of the species }
\item{Genus}{Genus of the species}
\item{Species}{Species Name}
\item{MDD_SciName}{Species binomial name}
...
}
}
}
\source{
Expert range maps of global mammal distributions harmonised to
three taxonomic authorities. \doi{doi:10.5281/zenodo.10806734}
}
\usage{
mddSpList_v1_3
}
\description{
A list of species related with shapefiles
in MDD. Includes taxonomy
}
\keyword{datasets}
